+++ FFXIVAPP.Plugin.Parse/Models/LinkedStats/MaxStat.cs
@@ -34,16 +34,16 @@ namespace FFXIVAPP.Plugin.Parse.Models.LinkedStats
 {
     public sealed class MaxStat : LinkedStat
     {
-        public MaxStat(string name, params Stat<decimal>[] dependencies) : base(name, 0m)
+        public MaxStat(string name, params Stat<double>[] dependencies) : base(name, 0)
         {
             AddDependency(dependencies[0]);
         }
 
-        public MaxStat(string name, decimal value) : base(name, 0m)
+        public MaxStat(string name, double value) : base(name, 0)
         {
         }
 
-        public MaxStat(string name) : base(name, 0m)
+        public MaxStat(string name) : base(name, 0)
         {
         }
 
@@ -54,8 +54,8 @@ namespace FFXIVAPP.Plugin.Parse.Models.LinkedStats
         /// <param name="newValue"> </param>
         public override void DoDependencyValueChanged(object sender, object previousValue, object newValue)
         {
-            var ovalue = (decimal) previousValue;
-            var nvalue = (decimal) newValue;
+            var ovalue = (double) previousValue;
+            var nvalue = (double) newValue;
             var delta = Math.Max(ovalue, nvalue) - Math.Min(ovalue, nvalue);
             if (delta > Value)
             {
